==> notebooks/02_feature_engg.ipynb <==
import pandas as pd
import os
df = pd.read_csv("../data/processed/training_data_p1.csv")
df["date_time"] = pd.to_datetime(df["date_time"])
df

df.dtypes

reg_targets = ["temperature", "precipitation"]
clf_target = ["weather_condition"]
feature_cols = [
    "humidity", "pressure", "wind_speed", "wind_direction", "cloud_coverage",
    "latitude", "longitude", "hour", "day", "weekday", "month"
]

from sklearn.model_selection import train_test_split

X = df[feature_cols]
y_reg = df[reg_targets]
y_clf = df[clf_target]

y_reg

y_clf

from sklearn.preprocessing import LabelEncoder

df_copy = df.copy()
le = LabelEncoder()
df["weather_condition_enc"] = le.fit_transform(df["weather_condition"])

weather_condition_mapping = dict(zip(le.classes_, le.transform(le.classes_)))

weather_condition_mapping

X_train_reg, X_test_reg, y_train_reg, y_test_reg = train_test_split(X, y_reg, test_size=0.2, random_state=42)

X_train_clf, X_test_clf, y_train_clf, y_test_clf = train_test_split(X, y_clf, test_size=0.2, random_state=42)

os.makedirs("../data/processed", exist_ok=True)

# Save regression splits
X_train_reg.to_csv("../data/processed/reg/X_train_reg.csv", index=False)
X_test_reg.to_csv("../data/processed/reg/X_test_reg.csv", index=False)
y_train_reg.to_csv("../data/processed/reg/y_train_reg.csv", index=False)
y_test_reg.to_csv("../data/processed/reg/y_test_reg.csv", index=False)

# Save classification splits
X_train_clf.to_csv("../data/processed/clf/X_train_clf.csv", index=False)
X_test_clf.to_csv("../data/processed/clf/X_test_clf.csv", index=False)
y_train_clf.to_csv("../data/processed/clf/y_train_clf.csv", index=False)
y_test_clf.to_csv("../data/processed/clf/y_test_clf.csv", index=False)

print("✅ All regression and classification splits saved under ../data/processed/")
